Heroku scheduler doesn't support monthly schedule, so run it daily and make the code be in charge of only running it once a month

+++ lib/tasks/delete_old_games.rake
@@ -6,6 +6,8 @@ namespace :delete do
 
   desc 'Delete all games'
   task :all_games => :environment do
-    Game.destroy_all
+    if(Date.today.day == 1)
+      Game.destroy_all
+    end
   end
 end
